ДОбавляет в статистику ранг пользователя
